Move DTOs back to api app folder, add controller API tags
