+++ apps/api/src/app/auth/dto/register.dto.ts
