+++ apps/api/src/app/cards/validator/card.validator.ts
