+++ apps/api/src/app/media/media.controller.ts
@@ -20,14 +20,16 @@ import { GetObjectCommandOutput } from "@aws-sdk/client-s3";
 import * as fs from "fs";
 import * as path from "path";
 import { S3 } from "@aws-sdk/client-s3";
-import { SetIdAndFileParam } from "@scholarsome/shared";
 // needed for multer file type declaration
 // eslint-disable-next-line @typescript-eslint/no-unused-vars, no-unused-vars
 import { Multer } from "multer";
 import { FileInterceptor } from "@nestjs/platform-express";
 import { AuthenticatedGuard } from "../auth/authenticated.guard";
 import * as sharp from "sharp";
+import { ApiTags } from "@nestjs/swagger";
+import { SetIdAndFileParam } from "./param/setIdAndFile.param";
 
+@ApiTags("Media")
 @Controller("media")
 export class MediaController {
   constructor(
