+++ apps/api/src/app/sets/dto/createSet.dto.ts
@@ -8,7 +8,7 @@ import {
   ValidateNested
 } from "class-validator";
 import { Type } from "class-transformer";
-import { CardValidator } from "../cards/cardValidator.validator";
+import { CardValidator } from "../../cards/validator/card.validator";
 
 export class CreateSetDto {
   @IsString()
