+++ apps/api/src/app/sets/param/setIdParam.param.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty, IsString } from "class-validator";
+
+export class SetIdParam {
+  @IsString()
+  @IsNotEmpty()
+    setId: string;
+}
