+++ apps/api/src/app/sets/sets.controller.ts
@@ -17,15 +17,7 @@ import { AuthenticatedGuard } from "../auth/authenticated.guard";
 import { SetsService } from "./sets.service";
 import { UsersService } from "../users/users.service";
 import { Request as ExpressRequest, Express } from "express";
-import {
-  ApiResponse,
-  ApiResponseOptions,
-  AuthorIdParam,
-  CreateSetDto,
-  CreateSetFromApkgDto,
-  SetIdParam,
-  UpdateSetDto
-} from "@scholarsome/shared";
+import { ApiResponse, ApiResponseOptions } from "@scholarsome/shared";
 import { Set } from "@prisma/client";
 import { FileInterceptor } from "@nestjs/platform-express";
 // needed for multer file type declaration
@@ -34,7 +26,14 @@ import { Multer } from "multer";
 import * as crypto from "crypto";
 import { CardsService } from "../cards/cards.service";
 import { CardMedia } from "@prisma/client";
+import { ApiCreatedResponse, ApiOperation, ApiTags } from "@nestjs/swagger";
+import { AuthorIdParam } from "./param/authorIdParam.param";
+import { CreateSetFromApkgDto } from "./dto/createSetFromApkg.dto";
+import { CreateSetDto } from "./dto/createSet.dto";
+import { UpdateSetDto } from "./dto/updateSet.dto";
+import { SetIdParam } from "./param/setIdParam.param";
 
+@ApiTags("Sets")
 @Controller("sets")
 export class SetsController {
   /**
@@ -51,7 +50,12 @@ export class SetsController {
    *
    * @returns `Set` object
    */
+  @ApiOperation( {
+    summary: "Get a set",
+    description: "Gets a set given a set ID"
+  })
   @Get(":setId")
+  @ApiCreatedResponse({ type: Set })
   async set(@Param() params: SetIdParam, @Request() req: ExpressRequest): Promise<ApiResponse<Set>> {
     const set = await this.setsService.set({
       id: params.setId
