+++ libs/shared/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./lib/types";
-export * from "./lib/dto";
