--- libs/shared/src/lib/dto/auth/index.ts
@@ -1,3 +1,0 @@
-export * from "./login.dto";
-export * from "./register.dto";
-export * from "./reset-password.dto";
