--- libs/shared/src/lib/dto/cards/index.ts
@@ -1,4 +1,0 @@
-export * from "./cardIdParam.param";
-export * from "./createCardDto.dto";
-export * from "./updateCardDto.dto";
-// Validators are not exported -> pseudo-private
