--- libs/shared/src/lib/dto/index.ts
@@ -1,5 +1,0 @@
-export * from "./auth";
-export * from "./cards";
-export * from "./sets";
-export * from "./users";
-export * from "./media";
