--- libs/shared/src/lib/dto/media/index.ts
@@ -1,1 +1,0 @@
-export * from "./setIdAndFileParam.param";
