--- libs/shared/src/lib/dto/sets/index.ts
@@ -1,5 +1,0 @@
-export * from "./authorIdParam.param";
-export * from "./createSetDto.dto";
-export * from "./setIdParam.param";
-export * from "./updateSetDto.dto";
-export * from "./createSetFromApkgDto";
