--- libs/shared/src/lib/dto/sets/setIdParam.param.ts
@@ -1,6 +1,0 @@
-import { IsString } from "class-validator";
-
-export class SetIdParam {
-  @IsString()
-    setId: string;
-}
