--- libs/shared/src/lib/dto/users/index.ts
@@ -1,1 +1,0 @@
-export * from "./userIdParam.param";
